Removed factory implementation for the selection strategies
* Moved selection strategies off dssd.py to selection_strategies.py file \n*Updated corresponding test files
